[PW-2846] Moving PM decorator logic to event subscriber
